requirements.txt
gopher-server==0.4.0
IMDbPY==6.8
lxml==4.5.0
pyfiglet==0.8.post1
SQLAlchemy==1.3.13
zope.interface==4.7.1
